+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 
+- ``0.1.x`` Titles
 - ``0.0.x`` Playground
 ---
+- ``0.1.3.1`` Creative work basic functionalities
+- ``0.1.2.1`` Basic tests for subtitle relationship
+- ``0.1.1.6`` Updated tests for titles to use DB entities
+- ``0.1.1.5`` Updated Title creation test to db entity
+- ``0.1.1.4`` Titles table column definitions
+- ``0.1.1.3`` Added start date to titles
+- ``0.1.1.2`` Base SQL table class (TODO: Title table)
+- ``0.1.1.1`` Basic titles definitions
 - ``0.0.5.2`` Features GSON
 - ``0.0.5.1`` Features default headers
 - ``0.0.4.1`` Post request & HTML forms